+++ packages/site/site-nextjs/src/sitePreview/appRouter/sitePreviewRoute.ts
@@ -5,7 +5,7 @@ import { cookies, draftMode } from "next/headers";
 import { redirect } from "next/navigation";
 import { type NextRequest, NextResponse } from "next/server";
 
-import { type PreviewParams, type SitePreviewParams, verifyJwt } from "../SitePreviewUtils";
+import { type PreviewParams, type SitePreviewParams, verifyJwt } from "../previewUtils";
 
 export async function sitePreviewRoute(request: NextRequest) {
     const params = request.nextUrl.searchParams;
@@ -33,26 +33,4 @@ export async function sitePreviewRoute(request: NextRequest) {
     draftMode().enable();
 
     return redirect(data.path);
-}
-
-/**
- * Helper for SitePreview
- * @param options.skipDraftModeCheck Allows skipping the draft mode check, only required when called from middleware.ts (see https://github.com/vercel/next.js/issues/52080)
- * @return If SitePreview is active the current preview settings
- */
-export async function previewParams(options: { skipDraftModeCheck: boolean } = { skipDraftModeCheck: false }): Promise<PreviewParams | null> {
-    if (!options.skipDraftModeCheck) {
-        if (!draftMode().isEnabled) return null;
-    }
-
-    const sitePreviewCookie = cookies().get("__comet_preview");
-    if (sitePreviewCookie) {
-        return verifyJwt<PreviewParams>(sitePreviewCookie.value);
-    }
-    const blockPreviewCookie = cookies().get("__comet_block_preview");
-    if (blockPreviewCookie) {
-        return verifyJwt<PreviewParams>(blockPreviewCookie.value);
-    }
-
-    return null;
 }
